+++ lib/cocktail/builder/i_calendar.ex
@@ -6,7 +6,7 @@ defmodule Cocktail.Builder.ICalendar do
   """
 
   alias Cocktail.{Rule, Schedule, Validation}
-  alias Cocktail.Validation.{Interval, Day, HourOfDay, MinuteOfHour, SecondOfMinute, TimeOfDay}
+  alias Cocktail.Validation.{Interval, Day, HourOfDay, MinuteOfHour, SecondOfMinute, TimeOfDay, TimeRange}
 
   @time_format_string "{YYYY}{0M}{0D}T{h24}{m}{s}"
 
@@ -92,7 +92,7 @@ defmodule Cocktail.Builder.ICalendar do
   @spec build_rule(Rule.t()) :: String.t()
   defp build_rule(%Rule{validations: validations_map, until: until, count: count}) do
     parts =
-      for key <- [:interval, :day, :hour_of_day, :minute_of_hour, :second_of_minute, :time_of_day],
+      for key <- [:interval, :day, :hour_of_day, :minute_of_hour, :second_of_minute, :time_of_day, :time_range],
           validation = validations_map[key],
           !is_nil(validation) do
         build_validation_part(key, validation)
@@ -110,6 +110,7 @@ defmodule Cocktail.Builder.ICalendar do
   defp build_validation_part(:minute_of_hour, %MinuteOfHour{minutes: minutes}), do: minutes |> build_minutes()
   defp build_validation_part(:second_of_minute, %SecondOfMinute{seconds: seconds}), do: seconds |> build_seconds()
   defp build_validation_part(:time_of_day, %TimeOfDay{times: times}), do: times |> build_times()
+  defp build_validation_part(:time_range, %TimeRange{} = time_range), do: time_range |> build_time_range()
 
   @spec build_until(Cocktail.time() | nil) :: [String.t()]
   defp build_until(nil), do: []
@@ -193,14 +194,26 @@ defmodule Cocktail.Builder.ICalendar do
     times_list =
       times
       |> Enum.sort()
-      |> Enum.map(fn {hour, min, sec} ->
-        hours = String.pad_leading("#{hour}", 2, "0")
-        mins = String.pad_leading("#{min}", 2, "0")
-        secs = String.pad_leading("#{sec}", 2, "0")
-        "#{hours}#{mins}#{secs}"
-      end)
+      |> Enum.map(&format_erl_time/1)
       |> Enum.join(",")
 
-    "BYTIME=#{times_list}"
+    "X-BYTIME=#{times_list}"
   end
+
+  defp format_erl_time({hour, min, sec}) do
+    hours = String.pad_leading("#{hour}", 2, "0")
+    mins = String.pad_leading("#{min}", 2, "0")
+    secs = String.pad_leading("#{sec}", 2, "0")
+
+    "#{hours}#{mins}#{secs}"
+  end
+
+  # "time range" validation
+
+  @spec build_time_range(TimeRange.t()) :: String.t()
+  defp build_time_range(%TimeRange{start_time: start_time, end_time: end_time, interval_seconds: interval}) do
+    "X-BYRANGE=" <> ([format_time(start_time), format_time(end_time), interval] |> Enum.join(","))
+  end
+
+  defp format_time(time), do: time |> Time.to_erl() |> format_erl_time
 end